+++ src/game/scenes/MainScene.ts
@@ -75,7 +75,7 @@ class MainScene extends Scene {
     this.moduleManager.register(Module.dpsIndicator, new DpsIndicatorModule(this, this.hero));
     this.moduleManager.register(Module.stageInfo, new StageInfoModule(this));
     this.moduleManager.register(Module.resourceList, new ResourceListModule(this, this.hero));
-    this.moduleManager.register(Module.activeEffects, new ActiveEffectsModule(this, this.hero));
+    this.moduleManager.register(Module.activeEffects, new ActiveEffectsModule(this));
     // cleanup any previous logs
     LogModule.cleanEntries();
     this.moduleManager.register(Module.logs, new LogModule(this));
@@ -107,25 +107,23 @@ class MainScene extends Scene {
 
     this.buildings = this.physics.add.group();
 
-    this.events.on("resume", () => {
-      this.hero.attackable.registerHealthRegenerationIfNecessary();
-    });
+    this.events.on("resume", () => this.onResume());
 
     // Listener for keyboard inputs
     this.input.keyboard?.on("keydown", (event: KeyboardEvent) => {
       if (event.key === "Escape") {
-        this.pauseCurrentScene();
+        this.onPause();
         this.scene.launch("PauseMenu");
       }
 
       if (event.key === "b") {
-        this.pauseCurrentScene();
+        this.onPause();
         // get disabled tiles and pass to build scene
         this.scene.launch("BuildMenuScene", {occupiedTiles: this.getOccupiedTiles()});
       }
 
       if (event.key === "c") {
-        this.pauseCurrentScene();
+        this.onPause();
         this.scene.launch("StatsScene", {statsManager: this.hero.stats});
       }
 
@@ -246,9 +244,15 @@ class MainScene extends Scene {
     });
   }
 
-  pauseCurrentScene() {
+  onPause() {
     this.scene.pause();
     this.hero.attackable.stopRegeneration();
+    this.events.emit("GamePaused");
+  }
+
+  onResume() {
+    this.hero.attackable.registerHealthRegenerationIfNecessary();
+    this.events.emit("GameResumed");
   }
 
   getOccupiedTiles() {